move asset_or_check_key_to_dep_node_handles to assets.py
branch-name: asset_or_check_key_to_dep_node_handles-mv
